requirements.txt
Flask==2.1.0
jinja2~=3.0.3
asgiref==3.3.4
numpy==1.22.4
Pillow==9.1.1
requests==2.28.0
tqdm==4.64.0
numba==0.56.4